+++ CHANGELOG.md
@@ -33,4 +33,5 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Server hardware
 - Server hardware type
 - Server profile
+- Server profile template
 - Uplink set